renamed ReaderInterestCardUiState to ReaderInterestsCardUiState
