+++ WordPress/src/main/java/org/wordpress/android/ui/reader/discover/viewholders/ReaderInterestCardViewHolder.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.reader_interest_card.*
 import org.wordpress.android.R
 import org.wordpress.android.ui.reader.discover.ReaderCardUiState
-import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderInterestCardUiState
+import org.wordpress.android.ui.reader.discover.ReaderCardUiState.ReaderInterestsCardUiState
 import org.wordpress.android.ui.reader.discover.ReaderInterestAdapter
 import org.wordpress.android.ui.utils.UiHelpers
 
@@ -15,7 +15,7 @@ class ReaderInterestCardViewHolder(
     parentView: ViewGroup
 ) : ReaderViewHolder(parentView, R.layout.reader_interest_card) {
     override fun onBind(uiState: ReaderCardUiState) {
-        uiState as ReaderInterestCardUiState
+        uiState as ReaderInterestsCardUiState
 
         if (interests_list.adapter == null) {
             interests_list.layoutManager = LinearLayoutManager(interests_list.context, RecyclerView.VERTICAL, false)
